16.06 12:36

+++ app/src/Controller/CommentController.php
@@ -6,6 +6,7 @@
 namespace App\Controller;
 
 use App\Entity\Comment;
+use App\Entity\User;
 use App\Form\Type\CommentType;
 use App\Service\CommentService;
 use Doctrine\ORM\OptimisticLockException;
@@ -56,6 +57,7 @@ class CommentController extends AbstractController
      * @return Response HTTP response
      */
     #[Route(name: "comment_index", methods: ["GET"])]
+//    #[IsGranted('VIEW', subject: 'comment')]
     public function index(Request $request): Response
     {
         $pagination = $this->commentService->getPaginatedList(
@@ -73,6 +75,7 @@ class CommentController extends AbstractController
      * @return Response HTTP response
      */
     #[Route("/{id}", name: "comment_show", methods: ["GET"],  requirements: ["id" => "[1-9]\d*"])]
+//    #[IsGranted('VIEW', subject: 'comment')]
     public function show(Comment $comment): Response
     {
         return $this->render(
@@ -92,8 +95,16 @@ class CommentController extends AbstractController
     #[Route("/create", methods: ["GET", "POST"], name: "comment_create")]
     public function create(Request $request): Response
     {
+        /** @var User $user */
+        $user = $this->getUser();
         $comment = new Comment();
-        $form = $this->createForm(CommentType::class, $comment, ['action' => $this->generateUrl('comment_create')]);
+        $comment->setAuthor($user);
+
+        $form = $this->createForm(
+            CommentType::class,
+            $comment,
+            ['action' => $this->generateUrl('comment_create')]
+        );
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -209,8 +220,17 @@ class CommentController extends AbstractController
      *
      */
     #[Route("/{id}/delete", methods: ["GET", "DELETE"], requirements: ["id" => "[1-9]\d*"], name: "comment_delete")]
+//    #[IsGranted('DELETE', subject: 'comment')]
     public function deleteComment(Request $request, Comment $comment): Response
     {
+        if ($comment->getAuthor() !== $this->getUser()) {
+            $this->addFlash(
+                'warning',
+                $this->translator->trans('message.record_not_found')
+            );
+            return $this->redirectToRoute('post_index');
+        }
+
         $form = $this->createForm(CommentType::class, $comment, [
             'method' => 'DELETE',
             'action' => $this->generateUrl('comment_delete', ['id' => $comment->getId()]),
